requirements.txt
# Core dependencies
beautifulsoup4==4.12.2
pyyaml==6.0.1
lxml==4.9.3

# CLI framework (future use)
click==8.1.7

# Web API
flask==3.0.0
flask-cors==4.0.0
PyGithub==2.1.1
python-dotenv==1.0.0
pytz==2024.1

# Testing
pytest==7.4.3
